weblog/requirements.txt
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
Django==4.1.2
djangorestframework==3.15.1
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.2.2
fonttools==4.39.4
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
MouseInfo==0.1.3
nest-asyncio==1.5.6
numpy==1.24.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.15.1
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2023.3
pywhatkit==5.4
pywin32==306
pyzmq==25.1.0
requests==2.28.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.3.2
traitlets==5.9.0
tzdata==2022.5
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.2.2
wikipedia==1.4.0
